refactor[#23] : 이메일 회원가입 폼을 저장하는 store 파일 생성
