+++ src/features/auth/store/index.ts
@@ -1,3 +1,4 @@
 export * from "./form";
 export * from "./petInformationForm";
 export * from "./region";
+export * from "./signUpByEmailForm";
