+++ src/shared/lib/form.ts
@@ -1,0 +1,8 @@
+export const useForm = ({ initialValues, onSubmit, validate }) => {};
+
+useForm({
+  initialValues: {
+    email: "",
+    passwoard: "",
+  },
+});
